Revert some parts of d_time for SAMx70
