+++ EVB-2/Makefile
@@ -173,7 +173,6 @@ C_SRCS += \
 	$(HW_LIBS_DIR)/drivers/SAMx70/d_i2c.c                                                                 \
 	$(HW_LIBS_DIR)/drivers/SAMx70/d_serial.c                                                            \
 	$(HW_LIBS_DIR)/drivers/SAMx70/d_time.c                                                                \
-	$(HW_LIBS_DIR)/drivers/common/d_time.c \
 	$(HW_LIBS_DIR)/bootloader/bootloaderApp.c                                                            \
 	$(HW_LIBS_DIR)/misc/rtos.c                                                                     \
 	$(HW_LIBS_DIR)/printf/printf.c                                                          \
